충돌해결
