+++ agaein_web/src/components/organism/createArticle/FormDate.tsx
@@ -1,7 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { FormRow, FormLabel } from '../../pages/createArticle/CreateArticle.style';
 import styled from 'styled-components';
 import Input from 'components/molecules/Input';
+
+import DatePicker from 'react-datepicker';
+import 'react-datepicker/dist/react-datepicker.css';
 
 const Form = styled.div`
     flex: 1;
@@ -10,10 +13,18 @@ const Form = styled.div`
 interface FormDateProps {}
 
 export function FormDate(props: FormDateProps) {
+    const [startDate, setStartDate] = useState<Date>(new Date());
+
     return (
         <FormRow>
             <FormLabel>실종일*</FormLabel>
-            <Form>{/* 캘린더 */}</Form>
+            <Form>
+              <DatePicker
+                dateFormat="yyyy/MM/dd"
+                selected={startDate}
+                onChange={(date: Date) => setStartDate(date)}
+              />
+            </Form>
         </FormRow>
     );
 }
